+++ src/richie/plugins/nesteditem/defaults.py
@@ -4,13 +4,10 @@ NestedItem plugin default settings
 from django.conf import settings
 from django.utils.translation import ugettext_lazy as _
 
+ACCORDION, LIST = "accordion", "list"
+
 NESTEDITEM_VARIANTS = getattr(
     settings,
     "RICHIE_NESTEDITEM_VARIANTS",
-    [
-        (None, _("Inherit")),
-        ("default", _("Default")),
-        ("accordion", _("Accordion")),
-        ("list", _("List")),
-    ],
+    [(LIST, _("List")), (ACCORDION, _("Accordion"))],
 )